Seperate arithmetic declaration and implementation -- for long term maintenance
